pyproject.toml
[tool.poetry]
name = "habit-tracker"
version = "0.1.0"
description = "Современный habit tracker с геймификацией"
authors = ["Inozemtsev Vladimir"]
license = "mit"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.2.5"
djangorestframework = "^3.16.1"
djangorestframework-simplejwt = "^5.5.1"
django-cors-headers = "^4.7.0"
psycopg2-binary = "^2.9.10"
python-decouple = "^3.8"
celery = "^5.5.3"
redis = "^6.4.0"
pillow = "^11.3.0"
django-filter = "^25.1"
django-extensions = "^4.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
